requirements.txt
fastapi
uvicorn
pymongo==4.3.3
certifi
python-dotenv
passlib
bcrypt
python-jose
python-multipart
motor==3.1.1
cryptography
cloudinary
python-multipart
pytest
httpx